refactor(utils): 아이템 관련 유틸함수 파일 이름 수정
